refactor(server): extract signal format/delta cluster → signal_format.mjs (scoreHistory as param)
